restore original tqdm.gui
